obj: fixed handling of the vertex color extension
previously if any mesh in an obj had vertex colors, every mesh after
that would also be assumed to have them, and end up reading outside of
the color array when assembling faces. Now we're keeping vertex position
and color data together in the same array, and make sure to only create
a color array if the current mesh has vertices with valid vertex colors.